+++ CHANGELOG.rst
@@ -1,6 +1,10 @@
 Change Log
 ==========
 
+* :+1: Added `current_user` method on the Client class to return the `User` object of the user connected to KE-chain.
+* :+1: Added various arguments to the `create_notification` method.
+* :+1: Added `edit` method to the Notification class.
+
 v3.6.2 (27MAR20)
 ----------------